backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
aiohttp>=3.9.0
gitpython>=3.1.0
PyYAML>=6.0
celery>=5.3.0
prometheus-client>=0.18.0
structlog>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
